requirements.txt
Pillow==11.2.1
pytesseract==0.3.13
schedule==1.2.2
selenium==4.32.0
